requirements.txt
beautifulsoup4
python-pptx
tqdm
requests